configurable qplot defaults
