+++ ext/QuanticaMakieExt/QuanticaMakieExt.jl
@@ -9,7 +9,7 @@ using Quantica: Lattice, LatticeSlice, AbstractHamiltonian, Hamiltonian,
       argerror, harmonics, sublats, siterange, site, norm,
       normalize, nsites, nzrange, rowvals, nonzeros, sanitize_SVector
 
-import Quantica: plotlattice, plotlattice!, plotbands, plotbands!, qplot, qplot!
+import Quantica: plotlattice, plotlattice!, plotbands, plotbands!, qplot, qplot!, qplotdefaults
 
 # Currying fallback
 Quantica.qplot(; kw...) = x -> Quantica.qplot(x; kw...)
@@ -18,6 +18,7 @@ Quantica.qplot!(; kw...) = x -> Quantica.qplot!(x; kw...)
 include("plotlattice.jl")
 include("plotbands.jl")
 include("tools.jl")
+include("defaults.jl")
 include("docstrings.jl")
 
 end # module
